Optimize Objective-C audio and video play

+++ FFmpegPlayer/player/Audio/FFAudioQueuePlayer.h
@@ -17,7 +17,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface FFAudioQueuePlayer : NSObject
 - (instancetype)initWithAudioInformation:(FFAudioInformation)audioInformation
                                 delegate:(id<FFAudioQueuePlayerDelegate>)delegate;
-- (void)receiveData:(uint8_t *)data length:(int)length aqBuffer:(AudioQueueBufferRef)aqBuffer;
+- (void)receiveData:(uint8_t *)data length:(int64_t)length aqBuffer:(AudioQueueBufferRef)aqBuffer;
 - (void)reuseAudioQueueBuffer:(AudioQueueBufferRef)aqBuffer;
 - (void)play;
 - (void)stop;
